Python/requirements.txt
fhir.resources==8.*
requests>=2.31.0 